refactor: Moved Traits to Traits folder and change namespaces
