+++ src/Http/Controllers/Apps/AppsController.php
@@ -15,9 +15,23 @@ use Kanvas\Http\Controllers\BaseController;
 use Kanvas\Apps\Apps\Actions\CreateAppsAction;
 use Kanvas\Apps\Apps\Actions\UpdateAppsAction;
 use Kanvas\Enums\HttpDefaults;
+use Kanvas\Users\Users\Models\Users;
 
 class AppsController extends BaseController
 {
+    protected Users $user;
+
+    /**
+    * Create a new controller instance.
+    *
+    * @param  UserRepository  $users
+    * @return void
+    */
+    public function __construct(Users $user)
+    {
+        $this->user = $user;
+    }
+
     /**
      * Fetch all apps
      *
@@ -35,6 +49,7 @@ class AppsController extends BaseController
             "current_page" => $response->currentPage(),
             "total" => $response->total()
         ];
+
         return response()->json($response);
     }
